Move Queuer to execution module.
* Break the cyclic association between WorkflowManager and Queuer.
* Move WorkflowManager._dispatcher_lock to execution module variable.
* Remove WorkflowManager._dispatcher_lock
* Remove Queuer._dispatcher_lock
* Remove RuntimeManager._dispatcher_lock

Ref #335
